updated renaming of pages
