+++ app/artist-page/personal-profile2/page.js
@@ -1,7 +1,7 @@
 import React from "react";
-import PersonalDetailsForm2 from "./_components/PersonalDetailsForm2";
+import PersonalDetailsForm2 from "./_components/Personal-Details-Form-2";
 import AppLayout from "../../../components/Applayout";
-import PersonalProfile2 from "./_components/PersonalProfile2";
+import PersonalProfile2 from "./_components/Personal-Profile-2";
 
 const page = () => {
   return (
