+++ app/landing-page/_components/Banner.js
