+++ app/landing-page/_components/Explore.js
