+++ app/landing-page/_components/Hero.js
